pyproject.toml
[tool.poetry]
name = "generative-agents"
version = "0.1.0"
description = ""
authors = ["Kuchiki <ykuchiki0827@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "0.27.0"
python-dotenv = "^1.0.1"
selenium = "4.8.2"
scipy = "^1.14.0"
pillow = "^10.4.0"
django = "2.2"
django-cors-headers = "2.5.3"
aiohttp = "3.8.3"
aiosignal = "1.3.1"
asgiref = "3.5.2"
async-generator = "1.10"
async-timeout = "4.0.2"
attrs = "22.2.0"
boto = "2.49.0"
botocore = "1.29.43"
click = "8.0.3"
cycler = "0.11.0"
dj-database-url = "0.5.0"
django-storages-redux = "1.3.3"
exceptiongroup = "1.1.0"
frozenlist = "1.3.3"
gensim = "3.8.0"
gunicorn = "20.1.0"
h11 = "0.14.0"
jmespath = "1.0.1"
joblib = ">=1.1.1"
kiwisolver = "1.4.4"
matplotlib = "3.7.2"
multidict = "6.0.4"
nltk = "3.6.5"
outcome = "1.2.0"
patsy = "0.5.3"
psycopg2-binary = "2.9.5"
pyparsing = "3.0.6"
pysocks = "1.7.1"
python-dateutil = "2.8.2"
pytz = "2021.3"
s3transfer = "0.6.0"
scikit-learn = "1.3.0"
scikit-posthocs = "0.7.0"
seaborn = "0.12.2"
six = "1.16.0"
sklearn = "0.0"
smart-open = "5.2.1"
sniffio = "1.3.0"
sortedcontainers = "2.4.0"
sqlparse = "0.4.3"
statsmodels = "0.13.5"
threadpoolctl = "3.0.0"
trio = "0.22.0"
trio-websocket = "0.9.2"
trueskill = "0.4.5"
wsproto = "1.2.0"
yarl = "1.8.2"
yellowbrick = "1.5"
numpy = "^2.0.0"
pandas = "^2.2.2"
google-cloud-translate = "^3.15.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
